Parallelism readme updates and bug fixes

+++ src/exchanges/bitstamp.cpp
@@ -130,7 +130,7 @@ double getActivePos(Parameters& params, std::string orderId) {
   unique_json root { authRequest(params, "/api/order_status/", options)};
   size_t arraySize = json_array_size(root.get());
   double sum = 0.0;
-  for (int i = 0; i < arraySize; ++i){
+  for (size_t i = 0; i < arraySize; ++i){
     //Is the key we are looking for btc in their return? we want the executed amount..
     sum += atof(json_string_value(json_object_get(json_array_get(root.get(), i),"btc")));
   }
